Broken into assemblies
