+++ Assets/Scripts/Details.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0632cb74004b7314f9b81e98f324e3fe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
