+++ Assets/Scripts/Editor/FloatReferenceDrawer.cs
@@ -1,3 +1,4 @@
+using ModularData;
 using UnityEditor;
 
 [CustomPropertyDrawer(typeof(FloatReference))]
