+++ Assets/Scripts/Hex.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee68d77ad52675d4787e795a4bff1723
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
