+++ Assets/Scripts/Hex/TTT.Hex.asmdef
@@ -1,0 +1,18 @@
+{
+    "name": "TTT.Hex",
+    "rootNamespace": "",
+    "references": [
+        "GUID:91ba5b220a43d5443ae04d679286f57d",
+        "GUID:1491147abca9d7d4bb7105af628b223e",
+        "GUID:be0903cd8e1546f498710afdc59db5eb"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
