+++ Assets/Scripts/Hex/TTT.Hex.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4c861275823ef884d91df8647f4caa34
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
