+++ Assets/Scripts/ModularData/ColorReference.cs
@@ -1,32 +1,35 @@
 using System;
 using UnityEngine;
 
-[Serializable]
-public class ColorReference
+namespace ModularData
 {
-    [Tooltip("Whether to use an inline (constant) value or an injected SO")]
-    public bool UseConstant = true;
-
-    /// <summary>
-    /// Defined by a constant/inline value from the inspector
-    /// </summary>
-    public Color Constant;
-
-    /// <summary>
-    /// Defined by a scriptable object injected by the inspector
-    /// </summary>
-    public ColorVariable Variable;
-
-    public ColorReference() { }
-
-    public ColorReference(Color value)
+    [Serializable]
+    public class ColorReference
     {
-        UseConstant = true;
-        Constant = value;
-    }
+        [Tooltip("Whether to use an inline (constant) value or an injected SO")]
+        public bool UseConstant = true;
 
-    /// <summary>
-    /// The actual color held by this object
-    /// </summary>
-    public Color Value => UseConstant ? Constant : Variable.Value;
+        /// <summary>
+        /// Defined by a constant/inline value from the inspector
+        /// </summary>
+        public Color Constant;
+
+        /// <summary>
+        /// Defined by a scriptable object injected by the inspector
+        /// </summary>
+        public ColorVariable Variable;
+
+        public ColorReference() { }
+
+        public ColorReference(Color value)
+        {
+            UseConstant = true;
+            Constant = value;
+        }
+
+        /// <summary>
+        /// The actual color held by this object
+        /// </summary>
+        public Color Value => UseConstant ? Constant : Variable.Value;
+    }
 }
