+++ Assets/Scripts/ModularData/ColorVariable.cs
@@ -1,11 +1,14 @@
 using UnityEngine;
 
-[CreateAssetMenu(
-    fileName = "ColorVariable",
-    menuName = "Scriptable Objects/Modular Data/ColorVariable"
-)]
-public class ColorVariable : ScriptableObject
+namespace ModularData
 {
-    [Tooltip("RGBA color")]
-    public Color Value;
+    [CreateAssetMenu(
+        fileName = "ColorVariable",
+        menuName = "Scriptable Objects/Modular Data/ColorVariable"
+    )]
+    public class ColorVariable : ScriptableObject
+    {
+        [Tooltip("RGBA color")]
+        public Color Value;
+    }
 }
