+++ Assets/Scripts/ModularData/FloatReference.cs
@@ -1,32 +1,35 @@
 using System;
 using UnityEngine;
 
-[Serializable]
-public class FloatReference
+namespace ModularData
 {
-    [Tooltip("Whether to use an inline (constant) value or an injected SO")]
-    public bool UseConstant = true;
-
-    /// <summary>
-    /// Defined by a constant/inline value from the inspector
-    /// </summary>
-    public float Constant;
-
-    /// <summary>
-    /// Defined by a scriptable object injected by the inspector
-    /// </summary>
-    public FloatVariable Variable;
-
-    public FloatReference() { }
-
-    public FloatReference(float value)
+    [Serializable]
+    public class FloatReference
     {
-        UseConstant = true;
-        Constant = value;
-    }
+        [Tooltip("Whether to use an inline (constant) value or an injected SO")]
+        public bool UseConstant = true;
 
-    /// <summary>
-    /// The actual float value held by this object
-    /// </summary>
-    public float Value => UseConstant ? Constant : Variable.Value;
+        /// <summary>
+        /// Defined by a constant/inline value from the inspector
+        /// </summary>
+        public float Constant;
+
+        /// <summary>
+        /// Defined by a scriptable object injected by the inspector
+        /// </summary>
+        public FloatVariable Variable;
+
+        public FloatReference() { }
+
+        public FloatReference(float value)
+        {
+            UseConstant = true;
+            Constant = value;
+        }
+
+        /// <summary>
+        /// The actual float value held by this object
+        /// </summary>
+        public float Value => UseConstant ? Constant : Variable.Value;
+    }
 }
