+++ Assets/Scripts/ModularData/FloatVariable.cs
@@ -1,10 +1,13 @@
 using UnityEngine;
 
-[CreateAssetMenu(
-    fileName = "FloatVariable",
-    menuName = "Scriptable Objects/Modular Data/FloatVariable"
-)]
-public class FloatVariable : ScriptableObject
+namespace ModularData
 {
-    public float Value;
+    [CreateAssetMenu(
+        fileName = "FloatVariable",
+        menuName = "Scriptable Objects/Modular Data/FloatVariable"
+    )]
+    public class FloatVariable : ScriptableObject
+    {
+        public float Value;
+    }
 }
