+++ Assets/Scripts/ModularData/TTT.ModularData.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "TTT.ModularData"
+}
