+++ Assets/Scripts/ModularData/TTT.ModularData.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 80a120755bd37c14f8c7866f38551991
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
