+++ Assets/Scripts/Terrain/TTT.Terrain.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "TTT.Terrain",
+    "rootNamespace": "",
+    "references": [
+        "GUID:80a120755bd37c14f8c7866f38551991"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
